# WIP: update GetUserStr - GetUserStrMd