- Old  GetUserStr

`goos: darwin
goarch: arm64
pkg: github.com/LightningTipBot/LightningTipBot/internal/telegram
BenchmarkGetUserName
BenchmarkGetUserName-10              	20432500	        58.77 ns/op
BenchmarkGetUserStrNoUsername
BenchmarkGetUserStrNoUsername-10     	13479517	        88.93 ns/op
BenchmarkGetUserStrNoFirstName
BenchmarkGetUserStrNoFirstName-10    	12947025	        92.27 ns/op
PASS`

- New  GetUserStr

`goos: darwin
goarch: arm64
pkg: github.com/LightningTipBot/LightningTipBot/internal/telegram
BenchmarkGetUserName
BenchmarkGetUserName-10              	19599570	        58.22 ns/op
BenchmarkGetUserStrNoUsername
BenchmarkGetUserStrNoUsername-10     	22363182	        53.96 ns/op
BenchmarkGetUserStrNoFirstName
BenchmarkGetUserStrNoFirstName-10    	21759808	        54.73 ns/op
PASS`


Both passing all defined tests 

`=== RUN   TestGetUserStr
=== RUN   TestGetUserStr/NoUserName
=== RUN   TestGetUserStr/UserName
=== RUN   TestGetUserStr/NoUserName#01
=== RUN   TestGetUserStr/NoUserName#02
--- PASS: TestGetUserStr (0.00s)
    --- PASS: TestGetUserStr/NoUserName (0.00s)
    --- PASS: TestGetUserStr/UserName (0.00s)
    --- PASS: TestGetUserStr/NoUserName#01 (0.00s)
    --- PASS: TestGetUserStr/NoUserName#02 (0.00s)
PASS`