@gohumble 2022-06-12
todo -- also update GetUserStrMd